convert storage module to ts
